Added Debugger functions to handle any unhandled exception.
Added code to the Thread class to subvert any exceptions to the Debugger, and also added it to the Windows main framework entry.

Need to add the functionality to Unix (including a segfault signal handler, thanks to Jarrett Billingsley for pointing out the relatively obvious).

I know stack unwinding will be hard. I don't think I will do it now.