Major changes.
Responder and Dispatcher have been added. This is the signal model that Djehuty will be based upon. Window, Control, etc have now become Responders and Timer, for example, is a simple Dispatcher. Responders are always also Dispatchers and unhandled signals go up the Responder tree. Responders have many Dispatchers but Dispatchers have only one Responder.

Tui and Gui spaces have been created for the various components of each. A TuiApplication class is for TUI apps and a GuiApplication class is for GUI apps. A TuiTetris game has been created as an example. type 'make tuitetris' to compile.

BaseWindow is dumb and has been removed. Much of the core has been cleaned up and simplified, yet there is more cleanup to do.

I will also have to outfit the networking helpers as Dispatchers. I need to think about some of the design of this in the next couple of days.