Reasonable success.
Here's the thing... you define loop unrolling as an operation

 (p) -> a -> (q) -> b -> (p)  ==>
 (p) -> a -> (q) -> b -> (r) -> a -> (q)

Ok. Things like a*(ab)* have to be analyzed before hand with a linear
pass. They will set a flag for the group so that it does not get added
if an 'a' (in this case) transition already exists.

Things like a*(ab)*a will allow the normal behavior by resetting all of
the flags.